fix(crash): platform-split unix code, harden test assertions, add key-value redaction test
Move collectCoreDumpDiagnostics/formatRlimitCore and
enableCrashForensics/logCoreDumpLimits/formatRlimit into
build-constrained _unix.go files with stubs for non-unix platforms.
Guard type assertions in TestWriteUnexpectedCloseReport to produce
clear failures instead of panics. Add key_value_params subtest for
secretKeyRegex coverage.
