+++ cmd/dumber/main_other.go
@@ -1,0 +1,9 @@
+//go:build !linux && !darwin
+
+package main
+
+import "context"
+
+func enableCrashForensics() {}
+
+func logCoreDumpLimits(_ context.Context) {}
