+++ internal/bootstrap/crash_report.go
@@ -16,7 +16,6 @@ import (
 	"time"
 
 	corelogging "github.com/bnema/dumber/internal/logging"
-	"golang.org/x/sys/unix"
 )
 
 const (
@@ -166,28 +165,6 @@ func currentReporter() unexpectedCloseReporter {
 		PID:         os.Getpid(),
 		PPID:        os.Getppid(),
 	}
-}
-
-func collectCoreDumpDiagnostics() unexpectedCloseCoreDump {
-	d := unexpectedCloseCoreDump{
-		Hint: "Run `coredumpctl list | rg -i \"dumber\"` and include matching entries in the issue.",
-	}
-	var limit unix.Rlimit
-	if err := unix.Getrlimit(unix.RLIMIT_CORE, &limit); err != nil {
-		d.RLimitCoreSoft = "unknown"
-		d.RLimitCoreHard = "unknown"
-		return d
-	}
-	d.RLimitCoreSoft = formatRlimitCore(limit.Cur)
-	d.RLimitCoreHard = formatRlimitCore(limit.Max)
-	return d
-}
-
-func formatRlimitCore(value uint64) string {
-	if value == unix.RLIM_INFINITY {
-		return "infinity"
-	}
-	return strconv.FormatUint(value, 10)
 }
 
 func readStartupProcessIDs(lockDir, sessionID string) (int, int) {
